requirements.txt
openai>=1.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
